PP-6361 Split charge_ft_tests.js into several files
Separate charge_ft_tests.js into different files for each endpoint:
card_details_ft_tests.js, confirm_ft_tests.js,
capture_waiting_ft_tests.js, cancel_ft_tests.js and the new,
slimmer charge_ft_tests.js.

No changes have been made to the tests: they’ve just been moved to
different files.

with @nimalank7
