+++ test/integration/charge_ft_tests.js
@@ -18,19 +18,17 @@ const {
   getChargeRequest,
   postChargeRequest,
   defaultConnectorResponseForGetCharge,
-  defaultAdminusersResponseForGetService,
-  connectorResponseForPutCharge
+  defaultAdminusersResponseForGetService
 } = helper
 const State = require('../../config/state')
 const serviceFixtures = require('../fixtures/service_fixtures')
-const random = require('../../app/utils/random')
 
 // Constants
 const app = proxyquire('../../server.js',
   {
     'memory-cache': {
-      get: function () { 
-        return false 
+      get: function () {
+        return false
       },
       '@global': true
     }
@@ -759,353 +757,6 @@ describe('chargeTests', function () {
           .expect(403)
           .end(done)
       })
-    })
-  })
-
-  describe('The /card_details/charge_id endpoint', function () {
-    it('It should show card details page if charge status is in "ENTERING CARD DETAILS" state', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      nock(process.env.CONNECTOR_HOST)
-        .put('/v1/frontend/charges/' + chargeId + '/status').reply(200)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge(enteringCardDetailsState, 'http://www.example.com/service', gatewayAccountId))
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      getChargeRequest(app, cookieValue, chargeId)
-        .expect(200)
-        .expect(function (res) {
-          const $ = cheerio.load(res.text)
-          expect($('#card-details #csrf').attr('value')).to.not.be.empty // eslint-disable-line
-          expect($('#amount').text()).to.eql('£23.45')
-          expect($('#payment-description').text()).to.contain('Payment Description')
-          expect($('#card-details').attr('action')).to.eql(frontendCardDetailsPostPath)
-          expect($('.withdrawal-text').text()).to.contains('Accepted credit and debit card types')
-          expect($('#address-country').length).to.equal(1)
-        })
-        .end(done)
-    })
-
-    it('It should show card details page with correct text for credit card only', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      nock(process.env.CONNECTOR_HOST)
-        .put('/v1/frontend/charges/' + chargeId + '/status').reply(200)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetChargeDebitCardOnly(enteringCardDetailsState, 'http://www.example.com/service', gatewayAccountId))
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      getChargeRequest(app, cookieValue, chargeId, '?debitOnly=true')
-        .expect(200)
-        .expect(function (res) {
-          const $ = cheerio.load(res.text)
-          expect($('.withdrawal-text').text()).to.contains('Credit card payments are not accepted. Please use a debit card.')
-        })
-        .end(done)
-    })
-
-    it('It should not show amex if it is excluded', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      nock(process.env.CONNECTOR_HOST)
-        .put('/v1/frontend/charges/' + chargeId + '/status').reply(200)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetChargeDebitCardOnly(enteringCardDetailsState, 'http://www.example.com/service', gatewayAccountId))
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      getChargeRequest(app, cookieValue, chargeId, '?removeAmex=true')
-        .expect(200)
-        .expect(function (res) {
-          expect(res.text).to.not.contain('american-express')
-        })
-        .end(done)
-    })
-
-    it('It should show 500 page if charge status cant be updated to "ENTERING CARD DETAILS" state with a 400 connector response', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      connectorResponseForPutCharge(chargeId, 400, { message: 'some error' })
-
-      getChargeRequest(app, cookieValue, chargeId)
-        .expect(500)
-        .end(done)
-    })
-
-    it('should fail to authorise when email patch fails', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      nock.cleanAll()
-
-      nock(process.env.CARDID_HOST)
-        .post('/v1/api/card', () => {
-          return true
-        })
-        .reply(200, { brand: 'visa', label: 'visa', type: 'D' })
-
-      nock(process.env.CONNECTOR_HOST)
-        .patch('/v1/frontend/charges/23144323')
-        .reply(500)
-      defaultConnectorResponseForGetCharge(chargeId, State.ENTERING_CARD_DETAILS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      connectorExpects(minimumConnectorCardData('5105105105105100'))
-        .reply(200)
-
-      postChargeRequest(app, cookieValue, minimumFormCardData('5105 1051 0510 5100'), chargeId, false)
-        .expect(500)
-        .end(done)
-    })
-
-    it('It should show 500 when email patch fails', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      nock(process.env.CONNECTOR_HOST)
-        .patch('/v1/frontend/charges/23144323')
-        .reply(500)
-
-      getChargeRequest(app, cookieValue, chargeId)
-        .expect(500)
-        .end(done)
-    })
-  })
-
-  describe('The /card_details/charge_id/confirm endpoint', function () {
-    beforeEach(function () {
-      nock.cleanAll()
-    })
-
-    it('should return the data needed for the UI when there is no corporate card information', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge('AUTHORISATION SUCCESS', 'http://www.example.com/service', gatewayAccountId))
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-      const cookieValue = cookie.create(chargeId)
-
-      getChargeRequest(app, cookieValue, chargeId, '/confirm')
-        .expect(200)
-        .expect(function (res) {
-          const $ = cheerio.load(res.text)
-          expect($('#confirmation #csrf').attr('value')).to.not.be.empty // eslint-disable-line
-          expect($('#card-number').text()).to.contains('●●●●●●●●●●●●1234')
-          expect($('#expiry-date').text()).to.contains('11/99')
-          expect($('#cardholder-name').text()).to.contains('Test User')
-          expect($('#address').text()).to.contains('line1, line2, city, postcode, United Kingdom')
-          expect($('#amount').text()).to.eql('£23.45')
-          expect($('#payment-description').text()).to.contain('Payment Description')
-          expect($('#payment-summary-breakdown-amount').length > 0).to.equal(false)
-          expect($('#payment-summary-corporate-card-fee').length > 0).to.equal(false)
-        })
-        .end(done)
-    })
-
-    it('should return the data needed for the UI when there is corporate card information', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetChargeCorporateCardOnly('AUTHORISATION SUCCESS', 'http://www.example.com/service', gatewayAccountId))
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-      const cookieValue = cookie.create(chargeId)
-
-      getChargeRequest(app, cookieValue, chargeId, '/confirm')
-        .expect(200)
-        .expect(function (res) {
-          const $ = cheerio.load(res.text)
-          expect($('#confirmation #csrf').attr('value')).to.not.be.empty // eslint-disable-line
-          expect($('#card-number').text()).to.contains('●●●●●●●●●●●●1234')
-          expect($('#expiry-date').text()).to.contains('11/99')
-          expect($('#cardholder-name').text()).to.contains('Test User')
-          expect($('#address').text()).to.contains('line1, line2, city, postcode, United Kingdom')
-          expect($('#amount').text()).to.eql('£25.95')
-          expect($('#payment-description').text()).to.contain('Payment Description')
-          expect($('#payment-summary-breakdown-amount').text()).to.contain('£23.45')
-          expect($('#payment-summary-corporate-card-fee').text()).to.contain('£2.50')
-        })
-        .end(done)
-    })
-
-    it('should post to the connector capture url looked up from the connector when a post arrives', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge(State.AUTH_SUCCESS, 'http://www.example.com/service', gatewayAccountId))
-        .post('/v1/frontend/charges/' + chargeId + '/capture').reply(204)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .set('Accept', 'application/json')
-        .expect(303, {})
-        .expect('Location', '/return/' + chargeId)
-        .end(done)
-    })
-
-    it('should error if no csrf token', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge(State.AUTH_SUCCESS, 'http://www.example.com/service', gatewayAccountId))
-        .post('/v1/frontend/charges/' + chargeId + '/capture').reply(200)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .set('Accept', 'application/json')
-        .expect(500)
-        .end(done)
-    })
-
-    it('connector failure when trying to capture should result in error page', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge(State.AUTH_SUCCESS, 'http://www.example.com/service', gatewayAccountId))
-        .post('/v1/frontend/charges/' + chargeId + '/capture').reply(500)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .set('Cookie', ['frontend_state=' + cookie.createWithReturnUrl(chargeId, undefined, 'http://www.example.com/service')])
-        .send({ csrfToken: helper.csrfToken() })
-        .expect(500)
-        .expect(function (res) {
-          const $ = cheerio.load(res.text)
-          expect($('#return-url').attr('href')).to.eql('/return/' + chargeId)
-        })
-        .end(done)
-    })
-
-    it('connector could not authorise capture results in error page', function (done) {
-      nock(process.env.CONNECTOR_HOST)
-        .get('/v1/frontend/charges/' + chargeId).reply(200, helper.rawSuccessfulGetCharge(State.AUTH_SUCCESS, 'http://www.example.com/service', gatewayAccountId))
-        .post('/v1/frontend/charges/' + chargeId + '/capture').reply(400)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .set('Cookie', ['frontend_state=' + cookie.createWithReturnUrl(chargeId, undefined, 'http://www.example.com/service')])
-        .send({ csrfToken: helper.csrfToken() })
-        .end(done)
-    })
-
-    it('should produce an error if the connector responds with a non 200', function (done) {
-      mockServer.get(connectorChargePath + chargeId).reply(404)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .set('Accept', 'application/json')
-        .expect(500)
-        .end(done)
-    })
-
-    it('should produce an error if the connector returns a non-200 status', function (done) {
-      defaultConnectorResponseForGetCharge(chargeId, State.AUTH_SUCCESS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-      mockServer.post(connectorChargePath + chargeId + '/capture').reply(1234)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .send({ csrfToken: helper.csrfToken() })
-        .expect(500)
-        .end(done)
-    })
-
-    it('should produce an error if the connector is unreachable for the confirm', function (done) {
-      defaultConnectorResponseForGetCharge(chargeId, State.AUTH_SUCCESS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .post(frontendCardDetailsPath + '/' + chargeId + '/confirm')
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .expect(500)
-        .end(done)
-    })
-  })
-
-  describe('capture waiting endpoint', function () {
-    it('should keep user in /capture_waiting when connector returns a capture ready state', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      defaultConnectorResponseForGetCharge(chargeId, State.CAPTURE_READY, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .get(frontendCardDetailsPath + '/' + chargeId + '/capture_waiting')
-        .set('Content-Type', 'application/x-www-form-urlencoded')
-        .set('Cookie', ['frontend_state=' + cookieValue])
-        .set('Accept', 'application/json')
-        .expect(200)
-        .end(done)
-    })
-
-    it('should take user to capture submitted view when charge in CAPTURE_SUBMITTED state', function (done) {
-      const cookieValue = cookie.create(chargeId)
-      defaultConnectorResponseForGetCharge(chargeId, State.CAPTURE_SUBMITTED, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      request(app)
-        .get(frontendCardDetailsPath + '/' + chargeId + '/capture_waiting')
-        .set('Content-Type', 'application/x-www-form-urlencoded')
-        .set('Cookie', ['frontend_state=' + cookieValue])
-        .set('Accept', 'application/json')
-        .expect(200)
-        .end(done)
-    })
-  })
-
-  describe('The cancel endpoint', function () {
-    it('should take user to cancel page on successful cancel when charge in entering card details state', function (done) {
-      const cancelEndpoint = frontendCardDetailsPath + '/' + chargeId + '/cancel'
-      defaultConnectorResponseForGetCharge(chargeId, State.ENTERING_CARD_DETAILS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      nock(process.env.CONNECTOR_HOST)
-        .post('/v1/frontend/charges/' + chargeId + '/cancel').reply(204)
-
-      request(app)
-        .post(cancelEndpoint)
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .expect(200)
-        .end(done)
-    })
-
-    it('should take user to cancel page on successful cancel when charge in authorisation successful state', function (done) {
-      const cancelEndpoint = frontendCardDetailsPath + '/' + chargeId + '/cancel'
-      defaultConnectorResponseForGetCharge(chargeId, State.AUTH_SUCCESS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      nock(process.env.CONNECTOR_HOST)
-        .post('/v1/frontend/charges/' + chargeId + '/cancel').reply(204)
-
-      request(app)
-        .post(cancelEndpoint)
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .expect(200)
-        .end(done)
-    })
-
-    it('should take user to error page on failed cancel', function (done) {
-      const cancelEndpoint = frontendCardDetailsPath + '/' + chargeId + '/cancel'
-      defaultConnectorResponseForGetCharge(chargeId, State.AUTH_SUCCESS, gatewayAccountId)
-      defaultAdminusersResponseForGetService(gatewayAccountId)
-
-      nock(process.env.CONNECTOR_HOST)
-        .post('/v1/frontend/charges/' + chargeId + '/cancel').reply(400)
-
-      request(app)
-        .post(cancelEndpoint)
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .expect(500)
-        .end(done)
-    })
-
-    it('should redirect user to service return_url on successful cancel when direct redirect enabled on the service', function (done) {
-      const returnUrl = 'http://www.example.com/service'
-      const gatewayAccountId = random.randomInt(1, 9999999).toString()
-      const cancelEndpoint = frontendCardDetailsPath + '/' + chargeId + '/cancel'
-      defaultConnectorResponseForGetCharge(chargeId, State.ENTERING_CARD_DETAILS, gatewayAccountId, returnUrl)
-      helper.adminusersResponseForGetServiceWithDirectRedirectEnabled(gatewayAccountId)
-
-      nock(process.env.CONNECTOR_HOST)
-        .post('/v1/frontend/charges/' + chargeId + '/cancel').reply(204)
-
-      request(app)
-        .post(cancelEndpoint)
-        .send({ csrfToken: helper.csrfToken() })
-        .set('Cookie', ['frontend_state=' + cookie.create(chargeId)])
-        .expect(302)
-        .expect(res => expect(res.headers.location).to.equal(returnUrl))
-        .end(done)
     })
   })
 })
